requirements.txt
Django==1.10
django-filter==0.14.0
django-nested-admin==3.0.8
djangorestframework==3.4.6
Markdown==2.6.6
python-monkey-business==1.0.0
six==1.10.0
